changed furnace code to round from float to int

+++ Scripts/resources.cs
@@ -3,5 +3,5 @@ using System;
 
 public partial class resources : Node
 {
-	public static float _money { get; set; } = 0f;
+	public static int _money { get; set; } = 0;
 }
